Reorganize analytics.ts into puhblic and private
