+++ packages/cli-kit/src/public/node/error-handler.ts
@@ -1,3 +1,5 @@
+import {reportEvent} from './analytics.js'
+import {getEnvironmentData} from '../../private/node/analytics.js'
 import {
   AbortSilent,
   CancelExecution,
@@ -7,7 +9,6 @@ import {
   cleanSingleStackTracePath,
 } from '../../error.js'
 import {debug, info} from '../../output.js'
-import {getEnvironmentData, reportEvent} from '../../analytics.js'
 import * as path from '../../path.js'
 import * as metadata from '../../metadata.js'
 import {fanoutHooks} from '../../plugins.js'
